requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
langchain==0.3.26
langchain-community==0.3.27
faiss-cpu==1.11.0
sentence-transformers==5.0.0
PyPDF2==3.0.1
pypdf==5.7.0
transformers==4.53.1
accelerate==1.8.1